requirements.txt
<<<<<<< HEAD
pymongo~=4.10.1
python-dotenv~=1.0.1
=======
pymongo~=4.10.1 
python-dotenv~=1.0.1
>>>>>>> nueva_rama
